requirements.txt
fastapi
pydantic
uvicorn[standard]
SQLAlchemy>=1.4
alembic
python-dotenv
asyncpg
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
psycopg2-binary
pyjwt
greenlet
python-jose[cryptography]